Bells of Pilgrimage
The Child of Prophecy has embarked on a journey to save the land of the Fae.
On her journey, she must ring 

N Bells of Pilgrimage.

The Child of Prophecy has an initial mana level of 

P. After that:

Each of the first 

X bells she rings will increase her mana level by 
10
10.
Each of the remaining bells will increase her mana level by 
5
5.
As a bonus, once the final bell is rung, her mana level will increase by a further 
20
20.

K bells have been rung so far. What is the current mana level of the Child of Prophecy?